feat(bench-plotter): refactor plotting pipeline and update CLI for sweep functionality
- Replaced the `generate_plots` CLI with a new `sweep` module for enhanced plotting capabilities.
- Removed deprecated files and functions related to the old plotting structure, streamlining the codebase.
- Updated the README to reflect changes in the CLI usage and the new output directory structure.
- Adjusted the pipeline to utilize `generate_plots_from_intervals`, improving the overall workflow and clarity.
- Enhanced tests to cover the new functionality and ensure robustness in the plotting process.
