--- bench-plotter/src/bench_plotter/generate_plots/__main__.py
@@ -1,3 +1,0 @@
-from bench_plotter.generate_plots import main
-
-main()
