+++ bench-plotter/src/bench_plotter/pipeline/__init__.py
@@ -1,5 +1,5 @@
 """Staged, parallel plotting pipeline: timing file -> plan -> fetch -> transform -> draw."""
 
-from .orchestrator import generate_plots_from_benchmark, generate_plots_from_intervals
+from .orchestrator import generate_plots_from_intervals
 
-__all__ = ["generate_plots_from_benchmark", "generate_plots_from_intervals"]
+__all__ = ["generate_plots_from_intervals"]
