+++ bench-plotter/src/bench_plotter/pipeline/orchestrator.py
@@ -1,7 +1,6 @@
-"""The single entry point: benchmark timing file -> plots, via four stages.
+"""The pipeline core: a list of intervals -> plots, via four stages.
 
-``generate_plots_from_benchmark`` reads ``benchmark_timing.json`` into an
-in-memory plan and drives the pipeline:
+``generate_plots_from_intervals`` drives the pipeline:
 
     1. plan      - interpret intervals + charts into typed PlotJobs
     2. fetch     - resolve every unique query concurrently (one event loop)
@@ -10,10 +9,7 @@ in-memory plan and drives the pipeline:
 
 This module owns none of the stage logic; it only wires the stages together and
 reports the outcome, so it stays small and each stage stays independently
-testable.
-
-``generate_plots_from_intervals`` is the reusable core used by the sweep module
-to plot one (tps, total_requests) configuration at a time.
+testable. The sweep module calls it once per (tps, total_requests) configuration.
 """
 
 from __future__ import annotations
@@ -27,35 +23,11 @@ matplotlib.use("Agg")
 from typing import Any, Dict, List
 
 from bench_plotter.metric_queries import get_charts_for_modes
-from bench_plotter.io_utils import load_json_data
 
 from .plan import build_plan
 from .fetch import fetch_all
 from .transform import transform_jobs
 from .draw import draw_all
-
-
-def _load_successful_intervals(path: str) -> List[Dict[str, Any]]:
-    """Load intervals, dropping any the benchmark recorded as non-success.
-
-    Accepts either a bare list of intervals (legacy) or an object with a
-    ``runs`` list (sweep format). Intervals without a ``status`` field are
-    kept (backward compatible).
-    """
-    data = load_json_data(path)
-    if isinstance(data, dict):
-        data = data.get("runs", [])
-    if not isinstance(data, list):
-        return []
-    kept = [
-        iv
-        for iv in data
-        if isinstance(iv, dict) and iv.get("status") in (None, "success")
-    ]
-    skipped = len(data) - len(kept)
-    if skipped:
-        print(f"Skipping {skipped} interval(s) with status != 'success'")
-    return kept
 
 
 def generate_plots_from_intervals(
@@ -96,26 +68,6 @@ def generate_plots_from_intervals(
     _report_failures(fetch_failures, draw_failures)
     print(f"Wrote {len(written)} plot(s) to '{output_dir}'")
     return written
-
-
-def generate_plots_from_benchmark(
-    test_intervals_path: str,
-    output_dir: str = "plots",
-    workers: int | None = None,
-    parallel: bool = True,
-) -> List[str]:
-    """Generate all plots for a benchmark run. Returns the written PNG paths.
-
-    Thin wrapper over :func:`generate_plots_from_intervals` that loads and
-    filters the timing file first.
-    """
-    intervals = _load_successful_intervals(test_intervals_path)
-    return generate_plots_from_intervals(
-        intervals,
-        output_dir=output_dir,
-        workers=workers,
-        parallel=parallel,
-    )
 
 
 def _report_failures(
