+++ bench-plotter/tests/test_prometheus_fetch.py
@@ -4,7 +4,7 @@ import asyncio
 import pytest
 from unittest.mock import AsyncMock, MagicMock, Mock, patch
 from bench_plotter.prometheus_fetch import (
-    range_step_for_window,
+    _step_for_range_seconds,
     query_range,
 )
 from bench_plotter.prometheus_matrix import (
@@ -12,26 +12,26 @@ from bench_plotter.prometheus_matrix import (
 )
 
 
-class TestRangeStepForWindow:
+class TestStepForRangeSeconds:
     def test_small_window(self) -> None:
         # Step is the 15s scrape_interval for any window under Prometheus's
         # per-series point cap.
-        assert range_step_for_window(300) == "15s"
+        assert _step_for_range_seconds(300) == "15s"
 
     def test_medium_window(self) -> None:
-        assert range_step_for_window(3600) == "15s"
+        assert _step_for_range_seconds(3600) == "15s"
 
     def test_large_window(self) -> None:
-        assert range_step_for_window(24 * 3600) == "15s"
+        assert _step_for_range_seconds(24 * 3600) == "15s"
 
     def test_window_at_point_cap_stays_15s(self) -> None:
         # 11_000 points * 15s == 165_000s: right at the cap, still 15s.
-        assert range_step_for_window(11_000 * 15) == "15s"
+        assert _step_for_range_seconds(11_000 * 15) == "15s"
 
     def test_window_over_point_cap_widens_step(self) -> None:
         # Beyond the cap the step widens just enough to stay under it, rather
         # than letting Prometheus reject the query with too many points.
-        assert range_step_for_window(48 * 3600) == "16s"
+        assert _step_for_range_seconds(48 * 3600) == "16s"
 
 
 class TestQueryRange:
